deleteOffer now returns an Offer object

+++ src/main/java/org/springframework/samples/petclinic/service/OfferService.java
@@ -44,8 +44,11 @@ public class OfferService {
 		}
 	}
 	
-	public void deleteOffer(int id) {
+	public Offer deleteOffer(int id) {
 		Offer offer = offerRepo.findById(id);
-		offerRepo.delete(offer);
+		if (offer != null) {
+			offerRepo.delete(offer);
+		}
+		return offer;
 	}
 }
